Add script for bumping Hermes version for a release
Summary:
React Native releases that bundle the Hermes source code will have a `sdks/.hermesversion` file which indicates the specific git tag used when pulling the Hermes source code.

This script provides a method for release coordinators to specify which Hermes tag will be used in a given release. The Hermes tag should already exist on the `facebook/hermes` repository on GitHub, although this script makes no determination as to its validity.

# Changelog

[Internal] Added new script for release coordinators

Reviewed By: cortinico

Differential Revision: D34460693

fbshipit-source-id: b2f882ba66d925034c3803aafe81de5204d9e33f